Keep scoped workflow draft names authoritative.
